+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 
 - "WhereWithOperatorsInput" merge to "WhereInput" and keep operators
 
+### Added
+
+- Add getters 'WhereInputType', 'CreateInputType', UpdateInputType from 'GrapheneDjangoCrud' object
+
 ## [1.0.3] - 2021-02-04
 
 ### Changed